+++ tencentcloud/cdn/v20180606/models.go
@@ -1416,6 +1416,7 @@ type BriefDomain struct {
 	// 加速服务状态
 	// rejected：域名审核未通过，域名备案过期/被注销导致
 	// processing：部署中
+	// closing：关闭中
 	// online：已启动
 	// offline：已关闭
 	Status *string `json:"Status,omitempty" name:"Status"`
@@ -6554,6 +6555,7 @@ type DetailDomain struct {
 	// 加速服务状态
 	// rejected：域名审核未通过，域名备案过期/被注销导致
 	// processing：部署中
+	// closing：关闭中
 	// online：已启动
 	// offline：已关闭
 	Status *string `json:"Status,omitempty" name:"Status"`