+++ tencentcloud/cvm/v20170312/models.go
@@ -49,7 +49,7 @@ type ActionTimer struct {
 	// 定时器名称，目前仅支持销毁一个值：TerminateInstances。
 	TimerAction *string `json:"TimerAction,omitempty" name:"TimerAction"`
 
-	// 执行时间，格式形如：2018-5-29 11:26:40,执行时间必须大于当前时间5分钟。
+	// 执行时间，按照ISO8601标准表示，并且使用UTC时间。格式为 YYYY-MM-DDThh:mm:ssZ。例如 2018-05-29T11:26:40Z，执行时间必须大于当前时间5分钟。
 	ActionTime *string `json:"ActionTime,omitempty" name:"ActionTime"`
 
 	// 扩展数据