+++ tencentcloud/iai/v20200303/client.go
@@ -4142,6 +4142,10 @@ func NewGetUpgradeGroupFaceModelVersionJobListResponse() (response *GetUpgradeGr
 }
 
 // GetUpgradeGroupFaceModelVersionJobList
+// 避免官网歧义
+//
+// 
+//
 // 获取人员库升级任务列表
 //
 // 可能返回的错误码:
@@ -4158,6 +4162,10 @@ func (c *Client) GetUpgradeGroupFaceModelVersionJobList(request *GetUpgradeGroup
 }
 
 // GetUpgradeGroupFaceModelVersionJobList
+// 避免官网歧义
+//
+// 
+//
 // 获取人员库升级任务列表
 //
 // 可能返回的错误码:
@@ -4204,6 +4212,10 @@ func NewGetUpgradeGroupFaceModelVersionResultResponse() (response *GetUpgradeGro
 }
 
 // GetUpgradeGroupFaceModelVersionResult
+// 避免官网歧义
+//
+// 
+//
 // 人员库升级结果查询
 //
 // 可能返回的错误码:
@@ -4221,6 +4233,10 @@ func (c *Client) GetUpgradeGroupFaceModelVersionResult(request *GetUpgradeGroupF
 }
 
 // GetUpgradeGroupFaceModelVersionResult
+// 避免官网歧义
+//
+// 
+//
 // 人员库升级结果查询
 //
 // 可能返回的错误码:
@@ -4798,6 +4814,10 @@ func NewRevertGroupFaceModelVersionResponse() (response *RevertGroupFaceModelVer
 }
 
 // RevertGroupFaceModelVersion
+// 同理
+//
+// 
+//
 // 本接口用于回滚人员库的人脸识别算法模型版本。单个人员库有且仅有一次回滚机会。
 //
 // 
@@ -4820,6 +4840,10 @@ func (c *Client) RevertGroupFaceModelVersion(request *RevertGroupFaceModelVersio
 }
 
 // RevertGroupFaceModelVersion
+// 同理
+//
+// 
+//
 // 本接口用于回滚人员库的人脸识别算法模型版本。单个人员库有且仅有一次回滚机会。
 //
 // 
@@ -5698,6 +5722,10 @@ func NewUpgradeGroupFaceModelVersionResponse() (response *UpgradeGroupFaceModelV
 }
 
 // UpgradeGroupFaceModelVersion
+// 避免官网歧义
+//
+// 
+//
 // 升级人员库。升级过程中，人员库仍然为原算法版本，人员库相关操作仍然支持。升级完成后，人员库为新算法版本。
 //
 // 单个人员库有且仅支持一次回滚操作。
@@ -5727,6 +5755,10 @@ func (c *Client) UpgradeGroupFaceModelVersion(request *UpgradeGroupFaceModelVers
 }
 
 // UpgradeGroupFaceModelVersion
+// 避免官网歧义
+//
+// 
+//
 // 升级人员库。升级过程中，人员库仍然为原算法版本，人员库相关操作仍然支持。升级完成后，人员库为新算法版本。
 //
 // 单个人员库有且仅支持一次回滚操作。