+++ tencentcloud/tse/v20201207/models.go
@@ -25,6 +25,9 @@ type ApolloEnvParam struct {
 	Name *string `json:"Name,omitempty" name:"Name"`
 
 	// 环境内引擎的节点规格 ID
+	// -1C2G
+	// -2C4G
+	// 兼容原spec-xxxxxx形式的规格ID
 	EngineResourceSpec *string `json:"EngineResourceSpec,omitempty" name:"EngineResourceSpec"`
 
 	// 环境内引擎的节点数量
@@ -38,6 +41,9 @@ type ApolloEnvParam struct {
 
 	// 子网 ID。在 VPC 的子网内分配一个 IP 作为 ConfigServer 的访问地址
 	SubnetId *string `json:"SubnetId,omitempty" name:"SubnetId"`
+
+	// 环境描述
+	EnvDesc *string `json:"EnvDesc,omitempty" name:"EnvDesc"`
 }
 
 type BoundK8SInfo struct {
@@ -80,12 +86,17 @@ type CreateEngineRequestParams struct {
 	// 
 	// 引擎各版本及可选择的规格、节点数说明：
 	// apollo - STANDARD版本
-	// 规格列表：spec-qcr53kf1t（1C2G）,spec-qdr53kf2w（2C4G）
+	// 规格列表：1C2G、2C4G、4C8G、8C16G、16C32G
 	// 节点数：1，2，3，4，5
 	// 
 	// eureka - STANDARD版本
-	// 规格列表：spec-qvj6k7t4q（1C2G）,spec-qcr53kfjt（2C4G）,spec-qvj6k7t4m（4G8G）,spec-qcr54kfjt（8C16G）,spec-qcr55kfjt（16C32G）
+	// 规格列表：1C2G、2C4G、4C8G、8C16G、16C32G
 	// 节点数：3，4，5
+	// 
+	// polarismesh - STANDARD版本
+	// 规格列表：NUM50、NUM100、NUM200、NUM500、NUM1000、NUM5000、NUM10000、NUM50000
+	// 
+	// 兼容原spec-xxxxxx形式的规格ID
 	EngineProductVersion *string `json:"EngineProductVersion,omitempty" name:"EngineProductVersion"`
 
 	// 引擎所在地域。参考值说明：
@@ -115,7 +126,7 @@ type CreateEngineRequestParams struct {
 
 	// 付费类型。参考值：
 	// - 0：后付费
-	// - 1：预付费
+	// - 1：预付费（接口暂不支持创建预付费实例）
 	TradeType *int64 `json:"TradeType,omitempty" name:"TradeType"`
 
 	// 引擎的节点规格 ID。参见EngineProductVersion字段说明
@@ -180,12 +191,17 @@ type CreateEngineRequest struct {
 	// 
 	// 引擎各版本及可选择的规格、节点数说明：
 	// apollo - STANDARD版本
-	// 规格列表：spec-qcr53kf1t（1C2G）,spec-qdr53kf2w（2C4G）
+	// 规格列表：1C2G、2C4G、4C8G、8C16G、16C32G
 	// 节点数：1，2，3，4，5
 	// 
 	// eureka - STANDARD版本
-	// 规格列表：spec-qvj6k7t4q（1C2G）,spec-qcr53kfjt（2C4G）,spec-qvj6k7t4m（4G8G）,spec-qcr54kfjt（8C16G）,spec-qcr55kfjt（16C32G）
+	// 规格列表：1C2G、2C4G、4C8G、8C16G、16C32G
 	// 节点数：3，4，5
+	// 
+	// polarismesh - STANDARD版本
+	// 规格列表：NUM50、NUM100、NUM200、NUM500、NUM1000、NUM5000、NUM10000、NUM50000
+	// 
+	// 兼容原spec-xxxxxx形式的规格ID
 	EngineProductVersion *string `json:"EngineProductVersion,omitempty" name:"EngineProductVersion"`
 
 	// 引擎所在地域。参考值说明：
@@ -215,7 +231,7 @@ type CreateEngineRequest struct {
 
 	// 付费类型。参考值：
 	// - 0：后付费
-	// - 1：预付费
+	// - 1：预付费（接口暂不支持创建预付费实例）
 	TradeType *int64 `json:"TradeType,omitempty" name:"TradeType"`
 
 	// 引擎的节点规格 ID。参见EngineProductVersion字段说明
@@ -1001,6 +1017,10 @@ type EnvAddressInfo struct {
 	// config内网访问地址
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	ConfigIntranetAddress *string `json:"ConfigIntranetAddress,omitempty" name:"ConfigIntranetAddress"`
+
+	// 是否开启config内网clb
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	EnableConfigIntranet *bool `json:"EnableConfigIntranet,omitempty" name:"EnableConfigIntranet"`
 }
 
 type EnvInfo struct {