+++ tencentcloud/vod/v20180717/models.go
@@ -7760,9 +7760,10 @@ type DescribeDailyMostPlayedStatRequestParams struct {
 	// Top 数据的统计指标，取值有：
 	// <li>Traffic：播放流量，按播放流量统计 Top100 的数据。</li>
 	// <li>PlayTimes：播放次数，按播放次数统计播放 Top100 的数据。</li>
+	// 默认值为Traffic。
 	Metric *string `json:"Metric,omitempty" name:"Metric"`
 
-	// 点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。
+	// <b>点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。</b>
 	SubAppId *uint64 `json:"SubAppId,omitempty" name:"SubAppId"`
 }
 
@@ -7778,9 +7779,10 @@ type DescribeDailyMostPlayedStatRequest struct {
 	// Top 数据的统计指标，取值有：
 	// <li>Traffic：播放流量，按播放流量统计 Top100 的数据。</li>
 	// <li>PlayTimes：播放次数，按播放次数统计播放 Top100 的数据。</li>
+	// 默认值为Traffic。
 	Metric *string `json:"Metric,omitempty" name:"Metric"`
 
-	// 点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。
+	// <b>点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。</b>
 	SubAppId *uint64 `json:"SubAppId,omitempty" name:"SubAppId"`
 }